requirements.txt
# Core dependencies
# torch>=2.5.1  # Installed via conda
# torchvision>=0.20.1  # Installed via conda
transformers>=4.47.0
accelerate>=1.2.0
python-dotenv>=1.0.1
tqdm-loggable>=0.2.0
datasets>=3.1.0

# Optional dependencies
# Only needed for specific platforms or features
# flash-attn>=2.5.5  # Uncomment if using CUDA for faster attention
